requirements.txt
absl-py==1.4.0
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
idna==3.4
keras==2.13.1
libclang==16.0.6
magic-filter==1.0.11
Markdown==3.4.4
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
typing_extensions==4.5.0
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
yarl==1.9.2
